remove scope creep from WidgetRenderer to WidgetManager
