+++ src/context/ContextProvider.tsx
@@ -39,6 +39,8 @@ const UIProvider = memo(function UIProvider({
     loadWidgets,
     snapshotEditModeMacros,
     restoreEditModeMacros,
+    baseGlobalMacros,
+    setMacroOverrides,
   } = useWidgetContext();
 
   const ui = useUIManager(
@@ -54,6 +56,8 @@ const UIProvider = memo(function UIProvider({
     loadWidgets,
     snapshotEditModeMacros,
     restoreEditModeMacros,
+    baseGlobalMacros,
+    setMacroOverrides,
   );
 
   return <UIContext.Provider value={ui}>{children}</UIContext.Provider>;
